feat: foldering with MV Architecture & adjusting Feedback
