+++ FiveGuyes/FiveGuyes/Sources/Views/CameraPickerView.swift
@@ -1,0 +1,60 @@
+//
+//  CameraPickerView.swift
+//  FiveGuyes
+//
+//  Created by Shim Hyeonhee on 10/10/24.
+//
+
+import PhotosUI
+import SwiftUI
+
+struct CameraPickerView: UIViewControllerRepresentable {
+    @Binding var selectedImage: UIImage?
+    @Environment(\.presentationMode) var presentationMode
+
+    var onImagePicked: (UIImage?) -> Void
+    var onCancel: () -> Void
+
+    func makeUIViewController(context: Context) -> UIImagePickerController {
+        let picker = UIImagePickerController()
+        picker.sourceType = .camera
+        picker.delegate = context.coordinator
+        return picker
+    }
+
+    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
+
+    class Delegate: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+        var onImagePicked: (UIImage?) -> Void
+        var onCancel: () -> Void
+        
+        init(onImagePicked: @escaping (UIImage?) -> Void, onCancel: @escaping () -> Void) {
+            self.onImagePicked = onImagePicked
+            self.onCancel = onCancel
+        }
+        
+        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
+            let image = info[.originalImage] as? UIImage
+            onImagePicked(image)
+            picker.dismiss(animated: true)
+        }
+
+        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+            onCancel()
+            picker.dismiss(animated: true)
+        }
+    }
+
+    func makeCoordinator() -> Delegate {
+        return Delegate(onImagePicked: { image in
+            self.selectedImage = image
+            self.presentationMode.wrappedValue.dismiss()
+        }, onCancel: {
+            self.presentationMode.wrappedValue.dismiss()
+        })
+    }
+}
+
+#Preview {
+    CameraPickerView( selectedImage: .constant(nil), onImagePicked: { _ in }, onCancel: {} )
+}
